feat: add downtime duration to all recovery notifications
