+++ server/notification-providers/notification-provider.js
@@ -194,6 +194,30 @@ class NotificationProvider {
         }
         return axiosConfig;
     }
+    /**
+     * Format duration as human-readable string (e.g., "1h 23m", "45m 30s")
+     * TODO: Update below to `Intl.DurationFormat("en", { style: "short" }).format(duration)` once we are on a newer node version
+     * @param {number} timeInSeconds The time in seconds to format a duration for
+     * @returns {string} The formatted duration
+     */
+    static formatDuration(timeInSeconds) {
+        const hours = Math.floor(timeInSeconds / 3600);
+        const minutes = Math.floor((timeInSeconds % 3600) / 60);
+        const seconds = timeInSeconds % 60;
+
+        const durationParts = [];
+        if (hours > 0) {
+            durationParts.push(`${hours}h`);
+        }
+        if (minutes > 0) {
+            durationParts.push(`${minutes}m`);
+        }
+        if (seconds > 0 && hours === 0) {
+            durationParts.push(`${seconds}s`);
+        }
+
+        return durationParts.length > 0 ? durationParts.join(" ") : "0s";
+    }
 }
 
 module.exports = NotificationProvider;
